requirements.txt
mcstatus
aiohttp
pillow
